Moved Scheme definition in its own module
